### 2017-01-15 b0fada4
Use inspect.signature where available based on deprecation of inspect.getargspec
