### 2017-01-17 #155 @kyleniemeyer
Great! Please take a look at the reviewer guidelines: http://joss.theoj.org/about#reviewer_guidelines
