### 2017-01-17 #165 @whedon
[REVIEW]: ODES: a high level interface to ODE and DAE solvers
**Submitting author:** bmcage (<a href="http://orcid.org/0000-0001-8383-8068">Benny Malengier</a>)
**Repository:** <a href="https://github.com/bmcage/odes/" target ="_blank">https://github.com/bmcage/odes/</a>
**Version:** v2.3.0

- [x] **Authorship:** Has the submitting author (bmcage) made major contributions to the software?
