### 2017-01-17 #155 @kyleniemeyer
OK @arghdos the actual review will take place over in #165, so I'm going to close this issue.
