### 2017-01-22 bmcage/odes#46 @arghdos
Community guidelines must be added to address JOSS requirements
There should be clear guidelines for third-parties wishing to:
-Contribute to the software
-Report issues or problems with the software
-Seek support

See for example this [Contributing.md](https://gist.github.com/PurpleBooth/b24679402957c63ec426)
