### 2017-01-22 bmcage/odes#47 @arghdos
The release version for JOSS needs a DOI
See this [github page](https://guides.github.com/activities/citable-code/) for a good guide to get a DOI via Zenodo.  The badge can also be added to the read-me if desired
