### 2017-01-22 #165 @arghdos
This is an important and useful piece of software, providing an easy to use interface to state-of-the art stiff ODE solvers with many examples, ample documentation and unit tests.

I recommend publishing in JOSS after a few minor revisions have been made:

1. The latest release on the target repo does not match the release number submitted to JOSS [issue #45](https://github.com/bmcage/odes/issues/45).  I suspect this is just a minor error, if @bmcage confirms I think we can update the submitted version to 2.2.1 (or to 2.3.0 on the target repo)
2. As a suggestion, I would put the link to the [readthedocs.io](http://scikits-odes.readthedocs.io/en/latest/?badge=latest) documentation more clearly in the read-me (under the Documentation section perhaps?).  It is currently present as a badge, but it took a bit of hunting for me to find it.
3. Community guidelines must be added to address the [JOSS requirement](http://joss.theoj.org/about) [issue #46](https://github.com/bmcage/odes/issues/46)
4. The release version (see above) needs a DOI, e.g. via [Zenodo](https://zenodo.org/), [issue #47](https://github.com/bmcage/odes/issues/47)
5.  As a suggestion, explicit use of the references in the paper.bib (which do appear to be comprehensive) could be made in either the README.md, the paper.md, the readthedocs or all of the above
