### 2017-01-22 #165 @kyleniemeyer
Thanks for your review @arghdos! I will clarify on point 4 (https://github.com/bmcage/odes/issues/47) that we only require the DOI at the end of the review process, after all necessary changes have been made.

@bmcage can you address the above points, and the associated issues @arghdos made in the software repo?
