### 2017-01-23 #165 @bmcage
@kyleniemeyer  I'm working against some deadlines at the moment. I'll look at the points at the end of the week, should not be a problem, otherwise next week
We planned a release, but that got delayed, which is the reason for the version mismatch.
