### 2017-02-17 b6f6447
PR #46, Community guidelines must be added to address JOSS requirements
