### 2017-02-17 bmcage/odes#46 @bmcage
fixed b6f6447
