### 2017-02-17 bmcage/odes#47 @bmcage
Zenodo release done, as this is different from before, the Zenodo release with DOI is v2.3.1
Procedures changed for the future so all releases will be now via github release and Zenodo
commit 279a8f0424d9d12f6f6f72506415892451729322
