### 2017-02-17 8c9f357
bugfix: pypi only allows tar or zip, not both
