### 2017-02-17 bmcage/odes#45 @bmcage
Due to issues with Zenado, release done is v2.3.1 instead of v2.3.0.
So JOSS submission version should be changed in v2.3.1 also. See releases in https://github.com/bmcage/odes/releases
