### 2017-02-17 8073daf
bugfix: added suggested link to readthedocs by JOSS review https://github.com/openjournals/joss-reviews/issues/165
