### 2017-02-17 #165 @bmcage
I believe to have fixed up all needed issues.
1. release v2.3.0 done in normal way. Updated procedures to do releases so Zenodo picks them up, then did release v2.3.1 So the version here would best be changed to v2.3.1
2. readthedocs is still barebones at the moment aimed at the most common use. Nevertheless, link is now added to documentation section of the readme
3. community guidelines added
4. a release 2.3.1 done according to Zenodo procedure to have it picked up. DOI was generated

For 5. I need to look at some examples of how references are done. Only a suggestion though
