### 2017-02-19 #165 @kyleniemeyer
Hi @bmcage: thanks for making those changes!

1. Regarding mentioning the references in the paper, I agree the references should be mentioned directly in the text. You can see examples of this in a few past JOSS papers:

- [r3d2](http://joss.theoj.org/papers/10.21105/joss.00016)
- [PyGBe](http://joss.theoj.org/papers/10.21105/joss.00043)

2. In the other text of the paper, I noticed some inconsistent capitalization in things like ODES (vs. odes, Odes). Proper capitalization should also be used with: Python, SciPy, Cython. And, "in stead" should be replaced with "instead".

3. I noticed some of the badges at the top of the README seem to be broken (see screenshot). Can you fix these?

![README badges broken](https://dsh.re/35a14)

4. I am hesitant to approve the paper when the documentation page is in the early stages. It looks as though all the necessary information is in the README, but it would be more accessible if presented with pages on the ReadTheDocs site. (I do see the API documentation there.) Putting all of the documentation in a single file makes it somewhat difficult to read, particularly without a table of contents.

Is this something you would be willing to do?
