### 2017-03-15 d51b2b6
Fix setup.py and README so that classifiers work
