### 2017-03-17 3db4e81
doc fix: this is https://github.com/bmcage/odes/pull/49 with also meta data
