+++ setup.cfg
@@ -1,0 +1,3 @@
+[metadata]
+license_file = LICENSE.txt
+description-file = README.md
