### 2017-03-17 #165 @bmcage
1-3 have been fixed, though a release to PyPI will be needed before the fix is visible in the tags.

For 4, @aragilar is working on cleaning up docstrings, so readthedocs comes out nicer. Start of using readthedocs was added recently and is a work in progress.
