### 2017-03-22 #165 @kyleniemeyer
@bmcage great! I think once that work is done with the docstrings, then we can accept.
