### 2017-05-31 092eef7
Add PEP 518 (i.e. pyproject.toml) support
