### 2017-06-01 #165 @aragilar
I've made some progress on 4, but the main blocker there is working out what to do with cython. I may end up splitting the docs between rtd for the main documentation, and using github pages or similar for  the API docs.
