### 2017-06-06 9d45f4a
Add check-manifest and check-readme to tox
