+++ setup.cfg
@@ -1,3 +1,8 @@
 [metadata]
 license_file = LICENSE.txt
 description-file = README.md
+
+[check-manifest]
+ignore =
+    .travis.yml
+    *.c
