+++ apidocs/requirements.txt
@@ -1,2 +1,2 @@
-sphinx>=1.3
-mock
+sphinx>=1.4
+sphinx_rtd_theme
