+++ scikits/odes/sundials/idas/idas.h
@@ -1,2106 +1,0 @@
-/*
- * -----------------------------------------------------------------
- * $Revision: 4845 $
- * $Date: 2016-08-03 15:45:09 -0700 (Wed, 03 Aug 2016) $
- * ----------------------------------------------------------------- 
- * Programmer(s): Radu Serban @ LLNL
- * -----------------------------------------------------------------
- * LLNS Copyright Start
- * Copyright (c) 2014, Lawrence Livermore National Security
- * This work was performed under the auspices of the U.S. Department 
- * of Energy by Lawrence Livermore National Laboratory in part under 
- * Contract W-7405-Eng-48 and in part under Contract DE-AC52-07NA27344.
- * Produced at the Lawrence Livermore National Laboratory.
- * All rights reserved.
- * For details, see the LICENSE file.
- * LLNS Copyright End
- * -----------------------------------------------------------------
- * This is the header (include) file for the main IDAS solver.
- * -----------------------------------------------------------------
- *
- * IDAS is used to solve numerically the initial value problem     
- * for the differential algebraic equation (DAE) system           
- *   F(t,y,y') = 0,                                               
- * given initial conditions                                       
- *   y(t0) = y0,   y'(t0) = yp0.                                  
- * Here y and F are vectors of length N.                          
- *
- * Additionally, IDAS can perform forward or adjoint sensitivity
- * analysis.
- * -----------------------------------------------------------------
- */
-
-#ifndef _IDAS_H
-#define _IDAS_H
-
-#include <stdio.h>
-#include <sundials/sundials_nvector.h>
-
-#ifdef __cplusplus     /* wrapper to enable C++ usage */
-extern "C" {
-#endif
-
-/* * =================================================================
- *              I D A S     C O N S T A N T S
- * =================================================================
- */
-
-/*
- * ----------------------------------------------------------------
- * Inputs to:
- *  IDAInit, IDAReInit, 
- *  IDASensInit, IDASensReInit, 
- *  IDAQuadInit, IDAQuadReInit,
- *  IDAQuadSensInit, IDAQuadSensReInit,
- *  IDACalcIC, IDASolve,
- *  IDAAdjInit
- * ----------------------------------------------------------------
- */
-
-/* itask */
-#define IDA_NORMAL           1
-#define IDA_ONE_STEP         2
-
-/* icopt */
-#define IDA_YA_YDP_INIT      1 
-#define IDA_Y_INIT           2
-
-/* ism */
-#define IDA_SIMULTANEOUS     1
-#define IDA_STAGGERED        2
-
-/* DQtype */
-#define IDA_CENTERED         1
-#define IDA_FORWARD          2
-
-/* interp */
-#define IDA_HERMITE          1
-#define IDA_POLYNOMIAL       2
-
-/*
- * ===============================================================
- * IDAS RETURN VALUES
- * ===============================================================
- */
-
-#define IDA_SUCCESS          0
-#define IDA_TSTOP_RETURN     1
-#define IDA_ROOT_RETURN      2
-
-#define IDA_WARNING          99
-
-#define IDA_TOO_MUCH_WORK   -1
-#define IDA_TOO_MUCH_ACC    -2
-#define IDA_ERR_FAIL        -3
-#define IDA_CONV_FAIL       -4
-
-#define IDA_LINIT_FAIL      -5
-#define IDA_LSETUP_FAIL     -6
-#define IDA_LSOLVE_FAIL     -7
[…]
