+++ scikits/odes/sundials/kinsol/kinsol_bbdpre.h
@@ -1,238 +1,0 @@
-/*
- * -----------------------------------------------------------------
- * $Revision: 4378 $
- * $Date: 2015-02-19 10:55:14 -0800 (Thu, 19 Feb 2015) $
- * -----------------------------------------------------------------
- * Programmer(s): Alan Hindmarsh, Radu Serban, and Aaron Collier @ LLNL
- * -----------------------------------------------------------------
- * LLNS Copyright Start
- * Copyright (c) 2014, Lawrence Livermore National Security
- * This work was performed under the auspices of the U.S. Department 
- * of Energy by Lawrence Livermore National Laboratory in part under 
- * Contract W-7405-Eng-48 and in part under Contract DE-AC52-07NA27344.
- * Produced at the Lawrence Livermore National Laboratory.
- * All rights reserved.
- * For details, see the LICENSE file.
- * LLNS Copyright End
- * -----------------------------------------------------------------
- * This is the header file for the KINBBDPRE module, for a
- * band-block-diagonal preconditioner, i.e. a block-diagonal
- * matrix with banded blocks, for use with KINSol, KINSp*,
- * and the parallel implementaion of the NVECTOR module.
- *
- * Summary:
- *
- * These routines provide a preconditioner matrix for KINSol that
- * is block-diagonal with banded blocks. The blocking corresponds
- * to the distribution of the dependent variable vector u amongst
- * the processes. Each preconditioner block is generated from
- * the Jacobian of the local part (associated with the current
- * process) of a given function g(u) approximating f(u). The blocks
- * are generated by each process via a difference quotient scheme,
- * utilizing the assumed banded structure with given half-bandwidths,
- * mudq and mldq. However, the banded Jacobian block kept by the
- * scheme has half-bandwidths mukeep and mlkeep, which may be smaller.
- *
- * The user's calling program should have the following form:
- *
- *   #include <sundials/sundials_types.h>
- *   #include <sundials/sundials_math.h>
- *   #include <sundials/sundials_iterative.h>
- *   #include <nvector_parallel.h>
- *   #include <kinsol.h>
- *   #include <kinsol/kinsol_bbdpre.h>
- *   ...
- *   MPI_Init(&argc,&argv);
- *   ...
- *   tmpl = N_VNew_Parallel(...);
- *   ...
- *   kin_mem = KINCreate();
- *   flag = KINInit(kin_mem,...,tmpl);
- *   ...
- *   flag = KINSptfqmr(kin_mem,...);
- *         -or-
- *   flag = KINSpbcg(kin_mem,...);
- *         -or-
- *   flag = KINSpgmr(kin_mem,...);
- *   ...
- *   flag = KINBBDPrecInit(kin_mem,...);
- *   ...
- *   KINSol(kin_mem,...);
- *   ...
- *   KINFree(&kin_mem);
- *   ...
- *   N_VDestroy_Parallel(tmpl);
- *   ...
- *   MPI_Finalize();
- *
- * The user-supplied routines required are:
- *
- *  func    the function f(u) defining the system to be solved:
- *          f(u) = 0
- *
- *  glocal  the function defining the approximation g(u) to f(u)
- *
- *  gcomm   the function to do necessary communication for glocal
- *
- * Notes:
- *
- * 1) This header file (kinsol_bbdpre.h) is included by the user for
- *    the definition of the KBBDData data type and for needed
- *    function prototypes.
- *
- * 2) The KINBBDPrecInit call includes half-bandwiths mudq and mldq
- *    to be used in the difference quotient calculation of the
- *    approximate Jacobian. They need not be the true half-bandwidths
- *    of the Jacobian of the local block of g, when smaller values may
- *    provide greater efficiency. Also, the half-bandwidths mukeep and
[…]

[…]
