+++ scikits/odes/sundials/arkode/arkode_direct.h
@@ -1,373 +1,0 @@
-/*---------------------------------------------------------------
- * Programmer(s): Daniel R. Reynolds @ SMU
- *---------------------------------------------------------------
- * LLNS/SMU Copyright Start
- * Copyright (c) 2015, Southern Methodist University and 
- * Lawrence Livermore National Security
- *
- * This work was performed under the auspices of the U.S. Department 
- * of Energy by Southern Methodist University and Lawrence Livermore 
- * National Laboratory under Contract DE-AC52-07NA27344.
- * Produced at Southern Methodist University and the Lawrence 
- * Livermore National Laboratory.
- *
- * All rights reserved.
- * For details, see the LICENSE file.
- * LLNS/SMU Copyright End
- *---------------------------------------------------------------
- * Common header file for the direct linear solvers in ARKODE.
- *--------------------------------------------------------------*/
-
-#ifndef _ARKDLS_H
-#define _ARKDLS_H
-
-#include <sundials/sundials_direct.h>
-#include <sundials/sundials_nvector.h>
-
-#ifdef __cplusplus  /* wrapper to enable C++ usage */
-extern "C" {
-#endif
-
-
-/*===============================================================
-  ARKDIRECT CONSTANTS
-===============================================================*/
-
-/* ARKDLS return values */
-#define ARKDLS_SUCCESS          0
-#define ARKDLS_MEM_NULL        -1
-#define ARKDLS_LMEM_NULL       -2
-#define ARKDLS_ILL_INPUT       -3
-#define ARKDLS_MEM_FAIL        -4
-#define ARKDLS_MASSMEM_NULL    -5
-
-/* Additional last_flag values */
-#define ARKDLS_JACFUNC_UNRECVR  -6
-#define ARKDLS_JACFUNC_RECVR    -7
-#define ARKDLS_MASSFUNC_UNRECVR -8
-#define ARKDLS_MASSFUNC_RECVR   -9
-
-
-/*===============================================================
-  FUNCTION TYPES
-===============================================================*/
-
-/*---------------------------------------------------------------
- Type: ARKDlsDenseJacFn
-
- A dense Jacobian approximation function Jac must be of type 
- ARKDlsDenseJacFn. Its parameters are:
-
- N   is the problem size.
-
- Jac is the dense matrix (of type DlsMat) that will be loaded
-     by a ARKDlsDenseJacFn with an approximation to the Jacobian 
-     matrix J = (df_i/dy_j) at the point (t,y). 
-
- t   is the current value of the independent variable.
-
- y   is the current value of the dependent variable vector,
-     namely the predicted value of y(t).
-
- fy  is the vector f(t,y).
-
- user_data is a pointer to user data - the same as the user_data
-     parameter passed to ARKodeSetFdata.
-
- tmp1, tmp2, and tmp3 are pointers to memory allocated for
- vectors of length N which can be used by a ARKDlsDenseJacFn
- as temporary storage or work space.
-
- A ARKDlsDenseJacFn should return 0 if successful, a positive 
- value if a recoverable error occurred, and a negative value if 
- an unrecoverable error occurred.
-
- NOTE: The following are two efficient ways to load a dense Jac:         
- (1) (with macros - no explicit data structure references)      
-     for (j=0; j < Neq; j++) {                                  
-       col_j = DENSE_COL(Jac,j);                                 
-       for (i=0; i < Neq; i++) {                                
-         generate J_ij = the (i,j)th Jacobian element           
-         col_j[i] = J_ij;                                       
-       }                                                        
-     }                                                          
[…]
