+++ scikits/odes/sundials/ida/ida_spils.h
@@ -1,326 +1,0 @@
-/*
- * -----------------------------------------------------------------
- * $Revision: 4378 $
- * $Date: 2015-02-19 10:55:14 -0800 (Thu, 19 Feb 2015) $
- * ----------------------------------------------------------------- 
- * Programmers: Alan Hindmarsh, Radu Serban and Aaron Collier @ LLNL
- * -----------------------------------------------------------------
- * LLNS Copyright Start
- * Copyright (c) 2014, Lawrence Livermore National Security
- * This work was performed under the auspices of the U.S. Department 
- * of Energy by Lawrence Livermore National Laboratory in part under 
- * Contract W-7405-Eng-48 and in part under Contract DE-AC52-07NA27344.
- * Produced at the Lawrence Livermore National Laboratory.
- * All rights reserved.
- * For details, see the LICENSE file.
- * LLNS Copyright End
- * -----------------------------------------------------------------
- * This is the common header file for the Scaled and Preconditioned
- * Iterative Linear Solvers in IDA.
- * -----------------------------------------------------------------
- */
-
-#ifndef _IDASPILS_H
-#define _IDASPILS_H
-
-#include <sundials/sundials_iterative.h>
-#include <sundials/sundials_nvector.h>
-
-#ifdef __cplusplus  /* wrapper to enable C++ usage */
-extern "C" {
-#endif
-
-/* 
- * -----------------------------------------------------------------
- * IDASPILS return values 
- * -----------------------------------------------------------------
- */
-
-#define IDASPILS_SUCCESS     0
-#define IDASPILS_MEM_NULL   -1 
-#define IDASPILS_LMEM_NULL  -2 
-#define IDASPILS_ILL_INPUT  -3
-#define IDASPILS_MEM_FAIL   -4
-#define IDASPILS_PMEM_NULL  -5
-
-/*
- * -----------------------------------------------------------------
- * Type : IDASpilsPrecSetupFn
- * -----------------------------------------------------------------
- * The optional user-supplied functions PrecSetup and PrecSolve
- * together must define the left preconditoner matrix P
- * approximating the system Jacobian matrix
- *    J = dF/dy + c_j*dF/dy'
- * (where the DAE system is F(t,y,y') = 0), and solve the linear
- * systems P z = r.   PrecSetup is to do any necessary setup
- * operations, and PrecSolve is to compute the solution of
- * P z = r.
- *
- * The preconditioner setup function PrecSetup is to evaluate and
- * preprocess any Jacobian-related data needed by the
- * preconditioner solve function PrecSolve.  This might include
- * forming a crude approximate Jacobian, and performing an LU
- * factorization on it.  This function will not be called in
- * advance of every call to PrecSolve, but instead will be called
- * only as often as necessary to achieve convergence within the
- * Newton iteration.  If the PrecSolve function needs no
- * preparation, the PrecSetup function can be NULL.
- *
- * Each call to the PrecSetup function is preceded by a call to
- * the system function res with the same (t,y,y') arguments.
- * Thus the PrecSetup function can use any auxiliary data that is
- * computed and saved by the res function and made accessible
- * to PrecSetup.
- *
- * A preconditioner setup function PrecSetup must have the
- * prototype given below.  Its parameters are as follows:
- *
- * tt  is the current value of the independent variable t.
- *
- * yy  is the current value of the dependent variable vector,
- *     namely the predicted value of y(t).
- *
- * yp  is the current value of the derivative vector y',
- *     namely the predicted value of y'(t).
- *
- * rr  is the current value of the residual vector F(t,y,y').
- *
- * c_j is the scalar in the system Jacobian, proportional to 1/hh.
- *
[…]

[…]
