### 2017-07-17 93b2603
Move C code in sundials_auxiliary.c to cython in common_defs.pyx
