### 2017-07-17 56bb268
Move C code in sundials_auxiliary.c to cython in common_defs.pyx
