### 2017-09-08 9dda52f
Use a generic float type instead of float
