+++ scikits/odes/sundials/cvode.pyx
@@ -15,13 +15,17 @@ from . import (
 
 from .c_sundials cimport realtype, N_Vector
 from .c_cvode cimport *
-from .common_defs cimport (nv_s2ndarray, ndarray2nv_s, ndarray2DlsMatd)
+from .common_defs cimport (
+    nv_s2ndarray, ndarray2nv_s, ndarray2DlsMatd, DTYPE_t,
+)
+from .common_defs import DTYPE # this is needed because we want DTYPE to be
+# accessible from python (not only in cython)
+
 
 # TODO: parallel implementation: N_VectorParallel
 # TODO: linsolvers: check the output value for errors
 # TODO: unify using float/double/realtype variable
 # TODO: optimize code for compiler
-
 
 SolverReturn = namedtuple(
     "SolverReturn", [
@@ -284,7 +288,7 @@ cdef int _jacdense(long int Neq, realtype tt,
         yy_tmp = aux_data.yy_tmp
         if aux_data.jac_tmp is None:
             N = np.alen(yy_tmp)
-            aux_data.jac_tmp = np.empty((N,N), float)
+            aux_data.jac_tmp = np.empty((N,N), DTYPE)
         jac_tmp = aux_data.jac_tmp
 
         nv_s2ndarray(yy, yy_tmp)
@@ -451,11 +455,11 @@ cdef int _prec_solvefn(realtype tt, N_Vector yy, N_Vector ff, N_Vector r, N_Vect
 
         if aux_data.r_tmp is None:
             N = np.alen(yy_tmp)
-            aux_data.r_tmp = np.empty(N, float)
+            aux_data.r_tmp = np.empty(N, DTYPE)
 
         if aux_data.z_tmp is None:
             N = np.alen(yy_tmp)
-            aux_data.z_tmp = np.empty(N, float)
+            aux_data.z_tmp = np.empty(N, DTYPE)
 
         r_tmp = aux_data.r_tmp
         z_tmp = aux_data.z_tmp
@@ -542,11 +546,11 @@ cdef int _jac_times_vecfn(N_Vector v, N_Vector Jv, realtype t, N_Vector y,
 
         if aux_data.r_tmp is None:
             N = np.alen(y_tmp)
-            aux_data.r_tmp = np.empty(N, float)
+            aux_data.r_tmp = np.empty(N, DTYPE)
 
         if aux_data.z_tmp is None:
             N = np.alen(y_tmp)
-            aux_data.z_tmp = np.empty(N, float)
+            aux_data.z_tmp = np.empty(N, DTYPE)
 
         v_tmp = aux_data.r_tmp
         Jv_tmp = aux_data.z_tmp
@@ -632,8 +636,8 @@ cdef class CV_data:
         self.user_data = None
         self.err_user_data = None
 
-        self.yy_tmp = np.empty(N, float)
-        self.yp_tmp = np.empty(N, float)
+        self.yy_tmp = np.empty(N, DTYPE)
+        self.yp_tmp = np.empty(N, DTYPE)
         self.jac_tmp = None
         self.g_tmp = None
         self.r_tmp = None
@@ -983,7 +987,7 @@ cdef class CVODE:
 
             # ...and to the auxiliary data object (holding runtime data)
             self.aux_data.rootfn = rootfn
-            self.aux_data.g_tmp  = np.empty([nr_rootfns,], float)
+            self.aux_data.g_tmp  = np.empty([nr_rootfns,], DTYPE)
 
             # TODO: Shouldn't be the rootn in the cvode obj unset first?
             flag = CVodeRootInit(cv_mem, nr_rootfns, _rootfn)
@@ -1030,7 +1034,7 @@ cdef class CVODE:
                 flag = CVodeSStolerances(cv_mem, <realtype> opts_rtol,
[…]
