+++ scikits/odes/sundials/ida.pyx
@@ -10,7 +10,11 @@ cimport numpy as np
 
 from .c_sundials cimport realtype, N_Vector
 from .c_ida cimport *
-from .common_defs cimport (nv_s2ndarray, ndarray2nv_s, ndarray2DlsMatd)
+from .common_defs cimport (
+    nv_s2ndarray, ndarray2nv_s, ndarray2DlsMatd, DTYPE_t,
+)
+from .common_defs import DTYPE # this is needed because we want DTYPE to be
+# accessible from python (not only in cython)
 from . import (
     IDASolveFailed, IDASolveFoundRoot, IDASolveReachedTSTOP, _get_num_args,
 )
@@ -303,7 +307,7 @@ cdef int _jacdense(long int Neq, realtype tt, realtype cj,
         yp_tmp = aux_data.yp_tmp
         if aux_data.jac_tmp is None:
             N = np.alen(yy_tmp)
-            aux_data.jac_tmp = np.empty((N,N), float)
+            aux_data.jac_tmp = np.empty((N,N), DTYPE)
         jac_tmp = aux_data.jac_tmp
 
         nv_s2ndarray(yy, yy_tmp)
@@ -388,9 +392,9 @@ cdef class IDA_data:
         self.user_data = None
         self.err_user_data = None
 
-        self.yy_tmp = np.empty(N, float)
-        self.yp_tmp = np.empty(N, float)
-        self.residual_tmp = np.empty(N, float)
+        self.yy_tmp = np.empty(N, DTYPE)
+        self.yp_tmp = np.empty(N, DTYPE)
+        self.residual_tmp = np.empty(N, DTYPE)
         self.jac_tmp = None
         self.g_tmp = None
 
@@ -726,7 +730,7 @@ cdef class IDA:
                 self.options['rootfn'] = tmpfun
 
             self.aux_data.rootfn = rootfn
-            self.aux_data.g_tmp  = np.empty([nr_rootfns,], float)
+            self.aux_data.g_tmp  = np.empty([nr_rootfns,], DTYPE)
 
             flag = IDARootInit(ida_mem, nr_rootfns, _rootfn)
 
@@ -773,7 +777,7 @@ cdef class IDA:
                 flag = IDASStolerances(ida_mem, <realtype> opts_rtol,
                                                 <realtype> opts_atol)
             else:
-                np_atol = np.asarray(opts_atol)
+                np_atol = np.asarray(opts_atol, dtype=DTYPE)
                 if np.alen(np_atol) != self.N:
                     raise ValueError("Array length inconsistency: 'atol' "
                                      "lenght (%i) differs from problem size "
@@ -842,7 +846,7 @@ cdef class IDA:
             self.options['validate_flags'] = validate_flags
             self._validate_flags = options["validate_flags"]
 
-    def init_step(self, double t0, object y0, object yp0,
+    def init_step(self, DTYPE_t t0, object y0, object yp0,
                    np.ndarray y_ic0_retn = None,
                    np.ndarray yp_ic0_retn = None):
         """
@@ -874,11 +878,11 @@ cdef class IDA:
                 message= String with message in case of an error
         """
         cdef int flag
-        cdef float time
+        cdef DTYPE_t time
         cdef np.ndarray[DTYPE_t, ndim=1] np_y0
         cdef np.ndarray[DTYPE_t, ndim=1] np_yp0
-        np_y0  = np.asarray(y0, dtype=float)
-        np_yp0 = np.asarray(yp0, dtype=float)
+        np_y0  = np.asarray(y0, dtype=DTYPE)
+        np_yp0 = np.asarray(yp0, dtype=DTYPE)
 
[…]
