### 2017-09-08 46808fc
Use a generic float type instead of float
