+++ scikits/odes/sundials/cvode.pyx
@@ -22,6 +22,7 @@ from .common_defs cimport (nv_s2ndarray, ndarray2nv_s, ndarray2DlsMatd)
 # TODO: unify using float/double/realtype variable
 # TODO: optimize code for compiler
 
+DTYPE = np.float
 
 SolverReturn = namedtuple(
     "SolverReturn", [
@@ -284,7 +285,7 @@ cdef int _jacdense(long int Neq, realtype tt,
         yy_tmp = aux_data.yy_tmp
         if aux_data.jac_tmp is None:
             N = np.alen(yy_tmp)
-            aux_data.jac_tmp = np.empty((N,N), float)
+            aux_data.jac_tmp = np.empty((N,N), DTYPE)
         jac_tmp = aux_data.jac_tmp
 
         nv_s2ndarray(yy, yy_tmp)
@@ -451,11 +452,11 @@ cdef int _prec_solvefn(realtype tt, N_Vector yy, N_Vector ff, N_Vector r, N_Vect
 
         if aux_data.r_tmp is None:
             N = np.alen(yy_tmp)
-            aux_data.r_tmp = np.empty(N, float)
+            aux_data.r_tmp = np.empty(N, DTYPE)
 
         if aux_data.z_tmp is None:
             N = np.alen(yy_tmp)
-            aux_data.z_tmp = np.empty(N, float)
+            aux_data.z_tmp = np.empty(N, DTYPE)
 
         r_tmp = aux_data.r_tmp
         z_tmp = aux_data.z_tmp
@@ -542,11 +543,11 @@ cdef int _jac_times_vecfn(N_Vector v, N_Vector Jv, realtype t, N_Vector y,
 
         if aux_data.r_tmp is None:
             N = np.alen(y_tmp)
-            aux_data.r_tmp = np.empty(N, float)
+            aux_data.r_tmp = np.empty(N, DTYPE)
 
         if aux_data.z_tmp is None:
             N = np.alen(y_tmp)
-            aux_data.z_tmp = np.empty(N, float)
+            aux_data.z_tmp = np.empty(N, DTYPE)
 
         v_tmp = aux_data.r_tmp
         Jv_tmp = aux_data.z_tmp
@@ -632,8 +633,8 @@ cdef class CV_data:
         self.user_data = None
         self.err_user_data = None
 
-        self.yy_tmp = np.empty(N, float)
-        self.yp_tmp = np.empty(N, float)
+        self.yy_tmp = np.empty(N, DTYPE)
+        self.yp_tmp = np.empty(N, DTYPE)
         self.jac_tmp = None
         self.g_tmp = None
         self.r_tmp = None
@@ -983,7 +984,7 @@ cdef class CVODE:
 
             # ...and to the auxiliary data object (holding runtime data)
             self.aux_data.rootfn = rootfn
-            self.aux_data.g_tmp  = np.empty([nr_rootfns,], float)
+            self.aux_data.g_tmp  = np.empty([nr_rootfns,], DTYPE)
 
             # TODO: Shouldn't be the rootn in the cvode obj unset first?
             flag = CVodeRootInit(cv_mem, nr_rootfns, _rootfn)
@@ -1133,7 +1134,7 @@ cdef class CVODE:
         if self._old_api:
             return (flag, time)
         else:
-            y_retn  = np.empty(np.alen(np_y0), float)
+            y_retn  = np.empty(np.alen(np_y0), DTYPE)
             y_retn[:] = np_y0[:]
             soln = SolverReturn(
                 flag=flag,
@@ -1270,7 +1271,7 @@ cdef class CVODE:
         #we test if rfn call doesn't give errors due to bad coding, as
         #cvode will ignore errors, it only checks return value (0 or 1 for error)
         if isinstance(rfn, CV_WrapRhsFunction):
[…]
