+++ scikits/odes/sundials/ida.pyx
@@ -20,6 +20,8 @@ from . import (
 # TODO: unify using float/double/realtype variable
 # TODO: optimize code for compiler
 
+DTYPE = np.float
+
 SolverReturn = namedtuple(
     "SolverReturn", [
         "flag", "values", "errors", "roots",
@@ -303,7 +305,7 @@ cdef int _jacdense(long int Neq, realtype tt, realtype cj,
         yp_tmp = aux_data.yp_tmp
         if aux_data.jac_tmp is None:
             N = np.alen(yy_tmp)
-            aux_data.jac_tmp = np.empty((N,N), float)
+            aux_data.jac_tmp = np.empty((N,N), DTYPE)
         jac_tmp = aux_data.jac_tmp
 
         nv_s2ndarray(yy, yy_tmp)
@@ -388,9 +390,9 @@ cdef class IDA_data:
         self.user_data = None
         self.err_user_data = None
 
-        self.yy_tmp = np.empty(N, float)
-        self.yp_tmp = np.empty(N, float)
-        self.residual_tmp = np.empty(N, float)
+        self.yy_tmp = np.empty(N, DTYPE)
+        self.yp_tmp = np.empty(N, DTYPE)
+        self.residual_tmp = np.empty(N, DTYPE)
         self.jac_tmp = None
         self.g_tmp = None
 
@@ -726,7 +728,7 @@ cdef class IDA:
                 self.options['rootfn'] = tmpfun
 
             self.aux_data.rootfn = rootfn
-            self.aux_data.g_tmp  = np.empty([nr_rootfns,], float)
+            self.aux_data.g_tmp  = np.empty([nr_rootfns,], DTYPE)
 
             flag = IDARootInit(ida_mem, nr_rootfns, _rootfn)
 
@@ -874,11 +876,11 @@ cdef class IDA:
                 message= String with message in case of an error
         """
         cdef int flag
-        cdef float time
+        cdef DTYPE_t time
         cdef np.ndarray[DTYPE_t, ndim=1] np_y0
         cdef np.ndarray[DTYPE_t, ndim=1] np_yp0
-        np_y0  = np.asarray(y0, dtype=float)
-        np_yp0 = np.asarray(yp0, dtype=float)
+        np_y0  = np.asarray(y0, dtype=DTYPE)
+        np_yp0 = np.asarray(yp0, dtype=DTYPE)
 
         flag, time = self._init_step(t0, np_y0, np_yp0, y_ic0_retn, yp_ic0_retn)
 
@@ -903,8 +905,8 @@ cdef class IDA:
             if self._old_api:
                 return (True, time)
             else:
-                y_retn  = np.empty(np.alen(np_y0), float)
-                yp_retn = np.empty(np.alen(np_y0), float)
+                y_retn  = np.empty(np.alen(np_y0), DTYPE)
+                yp_retn = np.empty(np.alen(np_y0), DTYPE)
                 # if no init cond computed, the start values are values at t=0
                 y_retn[:] = np_y0[:]
                 yp_retn[:] = np_yp0[:]
@@ -1160,7 +1162,7 @@ cdef class IDA:
 
         if not constraints_type is None:
             if not constraints_idx is None:
-                constraints_vars = np.zeros(N, float)
+                constraints_vars = np.zeros(N, DTYPE)
                 for idx in range(constraints_idx):
                     constraints_vars[constraints_idx[idx]] = constraints_type[idx]
             else:
@@ -1169,7 +1171,7 @@ cdef class IDA:
                   '  ''constraints_type'' the constraints_idx has to be of the'\
[…]
