### 2017-09-08 0f34807
Switch to trusty on travis
