+++ .travis.yml
@@ -3,6 +3,8 @@ language: python
 notifications:
   email: false
 
+dist: trusty
+
 addons:
   apt:
     packages:
@@ -19,20 +21,26 @@ cache:
     - $HOME/sundials
 
 env:
-  - TOXENV=py27 SUNDIALS_VERSION='2.7.0'
-  - TOXENV=py33 SUNDIALS_VERSION='2.7.0'
-  - TOXENV=py34 SUNDIALS_VERSION='2.7.0'
   - TOXENV=check-manifest SUNDIALS_VERSION='2.7.0'
   - TOXENV=checkreadme SUNDIALS_VERSION='2.7.0'
-#  - TOXENV=py35
-# commented out because of https://github.com/travis-ci/travis-ci/issues/4794
 
 matrix:
   include:
-    # needed to work around https://github.com/travis-ci/travis-ci/issues/4794
+    - python: 2.7
+      env:
+      - TOXENV=py27 SUNDIALS_VERSION='2.7.0'
+    - python: 3.3
+      env:
+      - TOXENV=py33 SUNDIALS_VERSION='2.7.0'
+    - python: 3.4
+      env:
+      - TOXENV=py34 SUNDIALS_VERSION='2.7.0'
     - python: 3.5
       env:
       - TOXENV=py35 SUNDIALS_VERSION='2.7.0'
+    - python: 3.6
+      env:
+      - TOXENV=py36 SUNDIALS_VERSION='2.7.0'
 
 install:
   - source ci_support/ensure_sundials_installed.sh
