### 2017-09-09 62e4eaf
Cleaned up sundials setup.py in prep for supporting multiple floats
