### 2017-09-09 2c67222
Cleaned up sundials setup.py in prep for supporting multiple floats
