### 2017-09-10 54f1ca8
Allow using sundials with different precisions
