+++ scikits/odes/sundials/setup.py
@@ -2,7 +2,7 @@ import os
 from distutils.log import info
 
 from numpy.distutils.system_info import get_info
-from scikits.odes._build import cython
+from scikits.odes._build import cython, write_pxi
 
 BASE_PATH = os.path.abspath(os.path.dirname(__file__))
 
@@ -35,7 +35,7 @@ if SUNDIALS_LIBDIR or SUNDIALS_INCLUDEDIR:
 
 elif SUNDIALS_INST_PREFIX is not None:
     SUNDIALS_LIBRARY_DIRS.append(os.path.join(SUNDIALS_INST_PREFIX, "lib"))
-    INCL_DIRS_SUNDIALS.append(os.path.join(SUNDIALS_INST_PREFIX, "include"))
+    SUNDIALS_INCLUDE_DIRS.append(os.path.join(SUNDIALS_INST_PREFIX, "include"))
     info("SUNDIALS installation path set to `{}` via $SUNDIALS_INST.".format(
         SUNDIALS_INST_PREFIX))
 else:
@@ -92,7 +92,7 @@ except:
 
 
 def configuration(parent_package='',top_path=None):
-    from numpy.distutils.misc_util import Configuration
+    from numpy.distutils.misc_util import Configuration, get_cmd
     info("=============================================")
     info("parent package is %s" % parent_package)
     info("top path is %s" % top_path)
@@ -100,6 +100,7 @@ def configuration(parent_package='',top_path=None):
     config = Configuration('sundials', parent_package, top_path)
 
     if use_lapack:
+        # set the final paths/libs for individual solvers
         CVODE_LIBRARIES.extend(SUNDIALS_LIBRARIES)
         IDA_LIBRARIES.extend(SUNDIALS_LIBRARIES)
         CVODE_INCLUDE_DIRS.extend(SUNDIALS_INCLUDE_DIRS)
@@ -107,9 +108,46 @@ def configuration(parent_package='',top_path=None):
         CVODE_LIBRARY_DIRS.extend(SUNDIALS_LIBRARY_DIRS)
         IDA_LIBRARY_DIRS.extend(SUNDIALS_LIBRARY_DIRS)
 
+        def get_sundials_config_pxi(ext, build_dir):
+            """
+            Create pxi file containing some of sundials build config
+
+            Don't ask why this is a function, something crazy about
+            distutils/numpy not setting _setup_distribution at the right time or
+            something...
+            """
+            SUNDIALS_CONFIG_H = "sundials/sundials_config.h"
+            config_cmd = config.get_config_cmd()
+            if config_cmd.check_macro_true(
+                "SUNDIALS_DOUBLE_PRECISION", headers=[SUNDIALS_CONFIG_H],
+                include_dirs=SUNDIALS_INCLUDE_DIRS
+            ):
+                SUNDIALS_FLOAT_TYPE = '"double"'
+                info("Found sundials built with double precision.")
+            elif config_cmd.check_macro_true(
+                "SUNDIALS_SINGLE_PRECISION", headers=[SUNDIALS_CONFIG_H],
+                include_dirs=SUNDIALS_INCLUDE_DIRS
+            ):
+                SUNDIALS_FLOAT_TYPE = '"single"'
+                info("Found sundials built with single precision.")
+            elif config_cmd.check_macro_true(
+                "SUNDIALS_EXTENDED_PRECISION", headers=[SUNDIALS_CONFIG_H],
+                include_dirs=SUNDIALS_INCLUDE_DIRS
+            ):
+                SUNDIALS_FLOAT_TYPE = '"extended"'
+                info("Found sundials built with extended precision.")
+            else:
+                # fall back to double
+                SUNDIALS_FLOAT_TYPE = '"double"'
[…]
