### 2017-09-11 df9ea05
Removed completed TODOs
