+++ scikits/odes/sundials/ida.pyx
@@ -21,7 +21,6 @@ from . import (
 
 # TODO: parallel implementation: N_VectorParallel
 # TODO: linsolvers: check the output value for errors
-# TODO: unify using float/double/realtype variable
 # TODO: optimize code for compiler
 
 SolverReturn = namedtuple(
