### 2017-09-11 bcb2a3f
Removed completed TODOs and unneeded import
