+++ scikits/odes/sundials/ida.pyx
@@ -19,7 +19,6 @@ from . import (
 
 # TODO: parallel implementation: N_VectorParallel
 # TODO: linsolvers: check the output value for errors
-# TODO: unify using float/double/realtype variable
 # TODO: optimize code for compiler
 
 DTYPE = np.float
