### 2017-09-13 ce42427
More precision changes
