### 2017-09-18 #165 @kyleniemeyer
Hi @aragilar @bmcage just wanted to check in on the progress here.
