### 2017-09-18 #165 @bmcage
I merged the new sphinx doc system.  Some other features, not joss related, to be fleshed out before we do a new release, but all is ok to resolve this issue in my opinion.
