### 2017-09-26 6f6826a
Test different precisions on travis
