+++ ci_support/install_sundials.sh
@@ -4,6 +4,7 @@ set -ex
 SUNDIALS=sundials-"${SUNDIALS_VERSION:-'2.7.0'}"
 SUNDIALS_FILE=$SUNDIALS.tar.gz
 SUNDIALS_URL=http://computation.llnl.gov/projects/sundials-suite-nonlinear-differential-algebraic-equation-solvers/download/$SUNDIALS_FILE
+PRECISION="${SUNDIALS_PRECISION:-'double'}"
 
 wget "$SUNDIALS_URL"
 
@@ -12,5 +13,5 @@ tar -xzvf "$SUNDIALS_FILE"
 mkdir sundials_build
 
 cd sundials_build &&
-    cmake -DCMAKE_INSTALL_PREFIX=$SUNDIALS_DIR -DLAPACK_ENABLE=ON ../$SUNDIALS &&
+    cmake -DCMAKE_INSTALL_PREFIX=$SUNDIALS_DIR -DLAPACK_ENABLE=ON -DSUNDIALS_PRECISION="$PRECISION" ../$SUNDIALS &&
     make && make install
