### 2017-09-26 a633616
Don't install ARKODE, it breaks multiple precision
