+++ ci_support/install_sundials.sh
@@ -12,6 +12,8 @@ tar -xzvf "$SUNDIALS_FILE"
 
 mkdir sundials_build
 
+# ARKODE is disabled as it does not support multiple precisions (will be fixed
+# in next sundials release)
 cd sundials_build &&
-    cmake -DCMAKE_INSTALL_PREFIX=$SUNDIALS_DIR -DLAPACK_ENABLE=ON -DSUNDIALS_PRECISION="$PRECISION" ../$SUNDIALS &&
+    cmake -DCMAKE_INSTALL_PREFIX=$SUNDIALS_DIR -DLAPACK_ENABLE=ON -DSUNDIALS_PRECISION="$PRECISION" -DBUILD_ARKODE:BOOL=OFF ../$SUNDIALS &&
     make && make install
