### 2017-09-26 9fd2d70
Disable examples as they don't support multiple precision
