### 2017-09-26 ce1ffa0
Single precision type in numpy is single not float
