### 2017-09-27 91b32d1
Use DTYPE in tests where needed
