### 2017-09-27 e22117d
Allow sundials built with single to fail tests
