+++ .travis.yml
@@ -56,6 +56,14 @@ matrix:
       env:
       - TOXENV=py36 SUNDIALS_VERSION='2.7.0' SUNDIALS_PRECISION='extended'
 
+  allow_failures:
+    - python: 2.7
+      env:
+      - TOXENV=py27 SUNDIALS_VERSION='2.7.0' SUNDIALS_PRECISION='single'
+    - python: 3.6
+      env:
+      - TOXENV=py36 SUNDIALS_VERSION='2.7.0' SUNDIALS_PRECISION='single'
+
 install:
   - source ci_support/ensure_sundials_installed.sh
   - pip install -U pip setuptools virtualenv wheel
