### 2017-10-13 #165 @kyleniemeyer
Sorry for the delay on handling this...

@arghdos could you look and see if the changes made satisfy your concerns?
