### 2017-10-17 #165 @arghdos
@kyleniemeyer @bmcage It definitely looks better, I just have a few comments on broken / missing links and I think it'll be good:

-  First, the link to the [API level docs](http://scikits-odes.readthedocs.io/en/latest/guide.html#lower-level-interfaces) in the lower level interfaces section appears to be offline.

-  Similarly a number of links (presumably) to the API-docs seem to be missing, e.g.: `scikits.odes.odeint.odeint` in [the User Guide](http://scikits-odes.readthedocs.io/en/latest/guide.html#structure-of-odes-and-user-s-guide), same with the solver links, e.g. `scikits.odes.sundials.cvode` in [Choosing a Solver](http://scikits-odes.readthedocs.io/en/latest/solvers.html#choosing-a-solver)
