### 2017-10-18 #165 @aragilar
The links should work once I get [doctr](https://drdoctr.github.io/doctr/index.html) set up. I've got a branch where I'm working on it, but I need to work out how best to integrate it with travis, how multiple versions should work etc. I'll post a link to the PR when it's ready.
