### 2017-12-18 b45e7e6
Update docs after building Travis build 134 of aragilar/odes
The docs were built from the branch 'add-doctr' against the commit
530b9182e2b56aefc6676853e0f13bbddbae28c5.

The Travis build that generated this commit is at
https://travis-ci.org/aragilar/odes/jobs/318351617.

The doctr command that was run is

    /home/travis/build/aragilar/odes/.tox/doctr/bin/doctr deploy --no-require-master --build-tags --built-docs apidocs/ dev
