+++ dev/index.rst
@@ -1,0 +1,28 @@
+.. Odes documentation master file, created by
+   sphinx-quickstart on Tue Feb  2 13:13:32 2016.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to the ODES scikit API Documentation
+============================================
+
+The ODES scikit provides access to Ordinary Differential Equation (ODE) solvers and Differential Algebraic Equation (DAE) solvers.
+
+This is the generated API docs, see https://scikits-odes.readthedocs.io/en/latest/ for the main documentation and user guide.
+
+Contents:
+
+.. toctree::
+   :maxdepth: 2
+
+   ode
+   dae
+   api
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
+
