### 2017-12-19 6bcb182
Update docs after building Travis build 136 of aragilar/odes
The docs were built from the branch 'add-doctr' against the commit
85d32e25696b47d66d1ada88f111f122bbcb511f.

The Travis build that generated this commit is at
https://travis-ci.org/aragilar/odes/jobs/318369056.

The doctr command that was run is

    /home/travis/build/aragilar/odes/.tox/doctr/bin/doctr deploy --no-require-master --build-tags --built-docs apidocs/_build/html dev
