### 2017-12-19 e999ca7
Add index builder
