### 2018-01-13 3ae2966
Update docstrings to improve their rendering in sphinx
