+++ scikits/odes/ode.py
@@ -1,6 +1,8 @@
+# -*- coding: utf-8 -*-
 # Authors: B. Malengier based on ode.py
 r"""
 First-order ODE solver
+======================
 
 User-friendly interface to various numerical integrators for solving a
 system of first order ODEs with prescribed initial conditions:
@@ -33,10 +35,7 @@ class OdeBase(object):
     Parameters
     ----------
     Rfn : function
-        A function which computes the required derivatives. The signature should
-        be:
-            func(t, y, y_dot, *args, **kwargs)
-        Note that *args and **kwargs handling are solver dependent.
+        A function which computes the required derivatives. The signature should be ``func(t, y, y_dot, *args, **kwargs)``. Note that *args and **kwargs handling are solver dependent.
 
     options : mapping
         Additional options for initialization, solver dependent
@@ -70,21 +69,32 @@ class OdeBase(object):
 
         Returns
         -------
-         if old_api
-            flag   - indicating return status of the solver
-            t      - numpy array of times at which the computations were successful
-            y      - numpy array of values corresponding to times t (values of y[i, :] ~ t[i])
-            t_err  - float or None - if recoverable error occured (for example reached maximum
-                     number of allowed iterations), this is the time at which it happened
-            y_err  - numpy array of values corresponding to time t_err
-         if old_api False:
-            A named tuple, with entries:
-                flag   = An integer flag (StatusEnum)
-                values = Named tuple with entries t and y
-                errors = Named tuple with entries t and y
-                roots  = Named tuple with entries t and y
-                tstop  = Named tuple with entries t and y
-                message= String with message in case of an error
+        old_api is False : namedtuple
+            namedtuple with the following attributes
+
+            =========== ==========================================
+            Field       Meaning
+            =========== ==========================================
+            ``flag``    An integer flag (StatusEnum)
+            ``values``  Named tuple with entries t and y
+            ``errors``  Named tuple with entries t and y
+            ``roots``   Named tuple with entries t and y
+            ``tstop``   Named tuple with entries t and y
+            ``message`` String with message in case of an error
+            =========== ==========================================
+
+        old_api is True : tuple
+            tuple with the following elements in order
+
+            ========== ==========================================
+            Field      Meaning
+            ========== ==========================================
+            ``flag``   indicating return status of the solver
+            ``t``      numpy array of times at which the computations were successful
+            ``y``      numpy array of values corresponding to times t (values of y[i, :] ~ t[i])
+            ``t_err``  float or None - if recoverable error occured (for example reached maximum number of allowed iterations), this is the time at which it happened
+            ``y_err``  numpy array of values corresponding to time t_err
+            ========== ==========================================
         """
         raise NotImplementedError('all ODE solvers must implement this')
 
@@ -94,23 +104,36 @@ class OdeBase(object):
 
         Parameters
         ----------
-        t0     - initial time
-        y0     - initial condition for y (can be list or numpy array)
+        t0 : number
+            initial time
+        y0 : array
+            initial condition for y (can be list or numpy array)
 
         Returns
         -------
-         if old_api:
-            flag  - boolean status of the computation (successful or error occured)
-            t_out - inititial time
+        old_api is False : namedtuple
+            namedtuple with the following attributes
 
-         if old_api False:
-            A named tuple, with entries:
-                flag   = An integer flag (StatusEnum)
-                values = Named tuple with entries t and y
-                errors = Named tuple with entries t and y
-                roots  = Named tuple with entries t and y
-                tstop  = Named tuple with entries t and y
-                message= String with message in case of an error
+            =========== ==========================================
+            Field       Meaning
+            =========== ==========================================
+            ``flag``    An integer flag (StatusEnum)
[…]
