+++ scikits/odes/dae.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # Authors: B. Malengier based on ode.py
 r"""
 First-order DAE solver
@@ -13,14 +14,10 @@ algebraic system of first order ODEs with prescribed initial conditions:
 
     \frac{d y(t=0)}{dt}[i]  = yprime0[i],
 
-where::
+where :math:`i = 0, ..., len(y0) - 1`; :math:`A` is a (possibly singular) matrix
+of size :math:`i × i`; and :math:`f(t,y)` is a vector of size :math:`i` or more generally, equations of the form
 
-    i = 0, ..., len(y0) - 1
-    A is a (possibly singular) matrix of size i x i
-    f(t,y) is a vector of size i
-
-or more generally, equations of the form
-
+.. math::
     G(t,y,y') = 0
 
 """
@@ -66,33 +63,44 @@ class DaeBase(object):
 
         Parameters
         ----------
-            tspan - an list/array of times at which the computed value will be
-                    returned. Must contain the start time as first entry.
-            y0    - list/numpy array of initial values
-            yp0   - list/numpy array of initial values of derivatives
+            tspan : list/array
+                A list of times at which the computed value will be returned. Must contain the start time as first entry.
+            y0 : list/array
+                List of initial values
+            yp0 : list/array
+                List of initial values of derivatives
 
         Returns
         -------
-         if old_api:
-            flag   - indicating return status of the solver
-            t      - numpy array of times at which the computations were successful
-            y      - numpy array of values corresponding to times t (values of y[i, :] ~ t[i])
-            yp     - numpy array of derivatives corresponding to times t (values of yp[i, :] ~ t[i])
-            t_err  - float or None - if recoverable error occured (for example reached maximum
-                     number of allowed iterations), this is the time at which it happened
-            y_err  - numpy array of values corresponding to time t_err
-            yp_err - numpy array of derivatives corresponding to time t_err
+        old_api is False : namedtuple
+            namedtuple with the following attributes
 
-         if old_api False (ida solver):
-            A named tuple, with entries:
-                flag   = An integer flag (StatusEnumXXX)
-                values = Named tuple with entries array t and array y and array
-                            ydot. y will correspond to y_retn value and ydot to
-                            yp_retn!
-                errors = Named tuple with entries t and y and ydot of error
-                roots  = Named tuple with entries array t and array y and array ydot
-                tstop  = Named tuple with entries array t and array y and array ydot
-                message= String with message in case of an error
+            =========== ==========================================
+            Field       Meaning
+            =========== ==========================================
+            ``flag``    An integer flag (StatusEnumXXX)
+            ``values``  Named tuple with entries array t and array y and array ydot. y will correspond to y_retn value and ydot to yp_retn!
+            ``errors``  Named tuple with entries t and y and ydot of error
+            ``roots``   Named tuple with entries array t and array y and array ydot
+            ``tstop``   Named tuple with entries array t and array y and array ydot
+            ``message`` String with message in case of an error
+            =========== ==========================================
+
+        old_api is True : tuple
+            tuple with the following elements in order
+
+            ========== ==========================================
+            Field      Meaning
+            ========== ==========================================
+            ``flag``   indicating return status of the solver
+            ``t``      numpy array of times at which the computations were successful
+            ``y``      numpy array of values corresponding to times t (values of y[i, :] ~ t[i])
+            ``yp``     numpy array of derivatives corresponding to times t (values of yp[i, :] ~ t[i])
+            ``t_err``  float or None - if recoverable error occured (for example reached maximum number of allowed iterations), this is the time at which it happened
+            ``y_err``  numpy array of values corresponding to time t_err
+            ``yp_err`` numpy array of derivatives corresponding to time t_err
+            ========== ==========================================
+
         """
[…]
