### 2018-01-26 0cbc568
FIX: Typo in travis config
