### 2018-01-26 a410b9d
ENH: Test docs (api-docs are tested by doctr)
