### 2018-01-26 #165 @aragilar
The API docs are now online. See https://bmcage.github.io/odes/dev/ for the rendered docs on master.
