### 2018-01-29 #165 @kyleniemeyer
@aragilar thanks!

@arghdos do the changes address your remaining concerns about the docs?
