### 2018-01-29 #165 @arfon
@kyleniemeyer - looks like there are some issues with the affiliations in the [`paper.md`](https://raw.githubusercontent.com/bmcage/odes/master/paper/paper.md).
