### 2018-01-30 #165 @aragilar
@arfon @kyleniemeyer Doing some cross-referencing on github, I think I've got everyone's afflictions now, which should be fixed by https://github.com/bmcage/odes/pull/64, I'm just waiting on confirmation from everyone that their details are correct, I'll ping here when I get confirmation.
