### 2018-02-13 #165 @aragilar
@arfon @kyleniemeyer The affiliations are up to date, and @bmcage is going to do a new release with all the documentation changes. Do you want any changes to the paper or anything before we do the new release?
