### 2018-02-14 #165 @arfon
Looks like the image referenced at `https://github.com/bmcage/odes/blob/master/docs/ipython/Performance%20tests.png` isn't working here. @aragilar Could you take a look at that please? It's probably better to have the PNG right next to the `paper.md` file.
