### 2018-02-14 67bf52a
Update links to images, examples in paper
