+++ paper/paper.md
@@ -47,10 +47,10 @@ ODES is a scikit offering extra ODE/DAE solvers, as an extension to what is avai
 
 ODES offers a high level object oriented API to Differential Equation solving in python. The backbone of ODES is the [SUNDIALS package](http://computation.llnl.gov/projects/sundials) [@hindmarsh2005sundials], which offers ODE and DAE solvers with root finding, preconditioning, error control and more.
 ODES can be used in python 2.7 or 3.2-3.5, while for speed purposes the integrator can be a Cython function instead of a pure python function. Comparison between different methods for a test problem is given in following graph:
-![Performance graph](https://github.com/bmcage/odes/blob/master/docs/ipython/Performance%20tests.png)
-You can generate above graph via the [Performance notebook](https://github.com/bmcage/odes/blob/master/docs/ipython/Performance%20tests.ipynb).
+![Performance graph](./PerformanceTests.png)
+You can generate above graph via the [Performance notebook](https://github.com/bmcage/odes/blob/master/ipython_examples/Performance%20tests.ipynb).
 
-As ODES is a python package, much thought has been given on ease of use. On one hand, novices can use the *odeint* convenience function, see example use in [simple.py](https://github.com/bmcage/odes/blob/master/docs/src/examples/odeint/simple.py). On the other hand, an object oriented interface is available via the *ode* and *dae* objects. See the [Documentation](https://scikits-odes.readthedocs.io/en/latest/) for details.
+As ODES is a python package, much thought has been given on ease of use. On one hand, novices can use the *odeint* convenience function, see example use in [simple.py](https://github.com/bmcage/odes/blob/master/docs/examples/odeint/simple.py). On the other hand, an object oriented interface is available via the *ode* and *dae* objects. See the [Documentation](https://scikits-odes.readthedocs.io/en/latest/) for details.
 
 For user new to solving ODEs, we recommend reading through @Hairer1993, which
 contains useful advice to find sucessful solutions.
