### 2018-02-14 #165 @aragilar
I've fixed links which were pointing to the wrong places, and copied across the png (I was going to symlink it, but I'm not sure whether the joss paper builder would support that).
