### 2018-02-14 #165 @kyleniemeyer
@aragilar successful build!

I do see a few issues:
-  the in-text reference to and citation for SciPy says "Jones et al. 2001–2001--"
- please capitalize Python everywhere you use it
- near the end, "For user" -> "For users"
- The Hindmarsh et al. reference should include a DOI (assuming one is available)

Please correct those things, and then we should be ready to accept. Once you do that, please archive the final version of the software & associated materials, and provide the DOI here.
