### 2018-02-15 #165 @aragilar
Fixed (the odes textbook lacks a doi apparently, even though there are ebooks...). I don't have access to make a new release, so @bmcage will make a release when he has time, and do the upload to zenodo.
