+++ common.py
@@ -21,8 +21,8 @@ INSTALL_REQUIRES = ['scipy']
 
 MAJOR = 2
 MINOR = 3
-MICRO = 2
-DEV = False
+MICRO = 3
+DEV = True
 
 CLASSIFIERS = [
         'Development Status :: 5 - Production/Stable',
