+++ README.md
@@ -6,7 +6,7 @@
 [![Supported implementations](https://img.shields.io/pypi/implementation/scikits.odes.svg)](https://pypi.python.org/pypi/scikits.odes/)
 [![PyPI](https://img.shields.io/pypi/status/scikits.odes.svg)](https://pypi.python.org/pypi/scikits.odes/)
 
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.292988.svg)](https://doi.org/10.5281/zenodo.292988)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1183272.svg)](https://doi.org/10.5281/zenodo.1183272)
 
 
 
@@ -28,13 +28,13 @@ A simple example solving the Van der Pol oscillator is as follows:
     import matplotlib.pyplot as plt
     import numpy as np
     from scikits.odes import ode
-    
+
     t0, y0 = 1, np.array([0.5, 0.5])  # initial condition
     def van_der_pol(t, y, ydot):
         """ we create rhs equations for the problem"""
         ydot[0] = y[1]
         ydot[1] = 1000*(1.0-y[0]**2)*y[1]-y[0]
-    
+
     solution = ode('cvode', van_der_pol, old_api=False).solve(np.linspace(t0,500,200), y0)
     plt.plot(solution.values.t, solution.values.y[:,0], label='Van der Pol oscillator')
     plt.show()
