### 2018-02-22 e264642
Update docs after building Travis build 93 of bmcage/odes
The docs were built from the tag 'v2.3.2b' against the commit
87aeef43b8901ef61ea76b6ed0782196eecf0ab6.

The Travis build that generated this commit is at
https://travis-ci.org/bmcage/odes/jobs/344814575.

The doctr command that was run is

    /home/travis/doctr/bin/doctr deploy --no-require-master --build-tags --command '/home/travis/doctr/bin/python build_index.py' --built-docs apidocs/_build/html version-v2.3.2b
