### 2018-02-22 #165 @bmcage
Release done v2.3.2, with all changes included. https://zenodo.org/record/1183272
There were some internal server errors in Zenodo originally, so had to draft the release a second time on github as v2.3.2b to have Zenodo upload working. DOI not active yet at time of writing this: should be https://doi.org/10.5281/zenodo.1183272
