### 2018-02-23 #165 @arfon
@bmcage @aragilar - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00165 :zap: :rocket: :boom:
